+++ engine/platform/glfw/platform_glfw_interface.h
@@ -50,7 +50,7 @@ private:
 
     bool initialized_;
     bool running_;
-    double previous_time_;
+    float previous_time_;
 };
 
 } // namespace glfw
